+++ Hashids/Hashids.h
@@ -1,0 +1,19 @@
+//
+//  Hashids.h
+//  Hashids
+//
+//  Created by Chad Brady on 8/25/15.
+//  Copyright (c) 2015 BuzzFeed. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for Hashids.
+FOUNDATION_EXPORT double HashidsVersionNumber;
+
+//! Project version string for Hashids.
+FOUNDATION_EXPORT const unsigned char HashidsVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Hashids/PublicHeader.h>
+
+
