--- Tests/Podfile
@@ -1,11 +1,0 @@
-# Uncomment this line to define a global platform for your project
-platform :ios, '8.0'
-
-use_frameworks!
-
-target 'HashidsTests' do
-
-  pod 'hashids', :path => "./../"
-
-end
-
